requirements.txt
# 基础依赖
python-dotenv>=0.19.0
requests>=2.26.0
Pillow>=8.3.0

# 地图生成
folium>=0.12.0
geopy>=2.2.0

# 图像处理
opencv-python>=4.5.0
numpy>=1.21.0

# 文本处理
jieba>=0.42.1
pandas>=1.3.0